fix copy move

+++ trivia.py
@@ -3,6 +3,16 @@ import shutil
 import subprocess
 import hashlib
 
+def run(cmd):
+    print cmd
+    return os.system(cmd)
+
+def popen(cmd):
+    print cmd
+    p = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
+    out_msg, err_msg = p.communicate()
+    return p.returncode, out_msg, err_msg
+
 def abspath(path):
     return os.path.abspath(path)
 
@@ -45,18 +55,26 @@ def remove(path):
         else:
             print "[error]not file or dir"
     else:
-        print "[error]no such:" + path
+        print "[error]no such path:" + path
 
-def move(src, dst):
-    if not exists(dst):
-        mkdirs(dst)
-    shutil.move(src, dst)
-
+# src: file only, no dir
 def copy(src, dst):
+    if not exists(src) or not isfile(src):
+        print "[error]no such file:" + src
+        return
     dir_path = dirpath(dst)
     if not exists(dir_path):
         mkdirs(dir_path)
     shutil.copy2(src, dst)
+
+# move file or directory
+def move(src, dst):
+    if not exists(src):
+        print "[error]no such src:" + src
+        return
+    if not exists(dst):
+        mkdirs(dst)
+    shutil.move(src, dst)
 
 def files_filter(rootdir, func):
     if not exists(rootdir) or not isdir(rootdir):
@@ -71,16 +89,7 @@ def files_filter(rootdir, func):
                 filepaths.append(filepath)
     return filepaths
 
-def run(cmd):
-    print cmd
-    return os.system(cmd)
-
-def popen(cmd):
-    print cmd
-    p = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True)
-    out_msg, err_msg = p.communicate()
-    return p.returncode, out_msg, err_msg
-
+# extract("[hello]", "[", "]")
 def extract(full_string, pattern_front, pattern_end):
     index0 = full_string.find(pattern_front)
     if index0 == -1:
